feat: add burn/claim page
this first commit is a draft for the ongoing work on this page. for starters, i've updated the useTheme hook to be type-safe.

i also went on to create a global store to sync the wallet state, although this is just a PoC for the entire implementation, using local state proved inefficient, as the click of "connect wallet" on the main card did not update the button state in the TopBar component.

i believe that this sets up a ground work for how we can manage reactivity and sync UI state accorss the application with zustand.

the UI still needs to match what's on the design, which i'll work on, and also finalize the process of adding new design tokens (color codes, basically) from figma, so i can properly implement the dark/light theme styles.