feat(ui): a complete refactor of the page to match what's on Figma
i've also included more components to proper reflect these changes.

the "connect wallet" button is now extensible with a new Gradient component to mimick the gradient border we have on most of the UI components on Figma. i also included new design tokens for the dark mode.

moving on to the "light mode" implementation now.