feat: add light theme styles and refactor the theme hook to use a context provider
the previous approach of enabling theming in the UI was buggy as there were stale values, and the body of the page doesn't recieve the correct theme value when it is switched.

each component tries to call it on its own, which led to race conditions. when you switch to dark mode, some components' style would still retain light mode style (and vice versa, unless you do a "hard-reload"), which is problematic.

the solution here was to move the logic into a global store with React context, so all components, irrespective of where they are in the tree have access to the same theme context.